+++ Changelog.md
@@ -1,6 +1,13 @@
 # v0.11.26 unreleased
 
-* Drop support for ruby 2.7
+* [#1409](https://github.com/mbj/mutant/pull/1409)
+
+  Fix config file support for AST pattern ignores. Before this fix the AST ignore
+  patterns would only be picked up when specified via CLI.
+
+* [#1405](https://github.com/mbj/mutant/pull/1406)
+
+  Drop support for ruby 2.7.
 
 # v0.11.25 2023-10-18